+++ patches/src/main/kotlin/app/morphe/patches/googlephotos/misc/gms/GmsCoreSupportPatch.kt
@@ -5,6 +5,7 @@
 package app.morphe.patches.googlephotos.misc.gms
 
 import app.morphe.patches.googlephotos.misc.extension.sharedExtensionPatch
+import app.morphe.patches.googlephotos.misc.login.selectedAccountPatch
 import app.morphe.patches.shared.compat.AppCompatibilities
 import app.morphe.patches.googlephotos.misc.gms.Constants.MORPHE_PHOTOS_PACKAGE_NAME
 import app.morphe.patches.googlephotos.misc.gms.Constants.PHOTOS_PACKAGE_NAME
@@ -21,6 +22,7 @@ val gmsCoreSupportPatch = gmsCoreSupportPatch(
     extensionPatch = sharedExtensionPatch,
     gmsCoreSupportResourcePatchFactory = ::gmsCoreSupportResourcePatch,
 ) {
+    dependsOn(selectedAccountPatch)
     compatibleWith(AppCompatibilities.GOOGLE_PHOTOS)
 }